chore: clean up regression test
